search_service/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.61.2
gunicorn==20.1.0
orjson==3.4.1
pydantic[dotenv]==1.9.0
PyJWT==2.6.0
uvicorn==0.12.2
uvloop==0.14.0
fastapi-pagination==0.9.3
requests==2.27.1
